+++ lib/blueprinter/base.rb
@@ -1,5 +1,5 @@
 require_relative 'blueprinter_error'
-require_relative 'helpers/active_record_helpers'
+require_relative 'configuration'
 require_relative 'extractor'
 require_relative 'extractors/association_extractor'
 require_relative 'extractors/auto_extractor'
@@ -7,6 +7,7 @@ require_relative 'extractors/block_extractor'
 require_relative 'extractors/hash_extractor'
 require_relative 'extractors/public_send_extractor'
 require_relative 'field'
+require_relative 'helpers/active_record_helpers'
 require_relative 'view'
 require_relative 'view_collection'
 
@@ -25,6 +26,9 @@ module Blueprinter
     # @param name [Symbol] to rename the identifier key in the JSON
     #   output. Defaults to method given.
     # @param extractor [AssociationExtractor,AutoExtractor,BlockExtractor,HashExtractor,PublicSendExtractor]
+    # @yield [object, options] The object and the options passed to render are
+    #   also yielded to the block.
+    #
     #   Kind of extractor to use.
     #   Either define your own or use Blueprinter's premade extractors.
     #   Defaults to AutoExtractor
@@ -35,9 +39,22 @@ module Blueprinter
     #     # other code
     #   end
     #
+    # @example Passing a block to be evaluated as the value.
+    #   class UserBlueprint < Blueprinter::Base
+    #     identifier :uuid do |user, options|
+    #       options[:current_user].anonymize(user.uuid)
+    #     end
+    #   end
+    #
     # @return [Field] A Field object
-    def self.identifier(method, name: method, extractor: AutoExtractor)
-      view_collection[:identifier] << Field.new(method, name, extractor)
+    def self.identifier(method, name: method, extractor: AutoExtractor.new, &block)
+      view_collection[:identifier] << Field.new(
+        method,
+        name,
+        extractor,
+        self,
+        block: block,
+      )
     end
 
     # Specify a field or method name to be included for serialization.
@@ -53,8 +70,18 @@ module Blueprinter
     # @option options [Symbol] :name Use this to rename the method. Useful if
     #   if you want your JSON key named differently in the output than your
     #   object's field or method name.
-    # @yield [Object] The object passed to `render` is also passed to the
-    #   block.
+    # @option options [String] :datetime_format Format Date or DateTime object
+    #   with given strftime formatting
+    # @option options [Symbol,Proc] :if Specifies a method, proc or string to
+    #   call to determine if the field should be included (e.g.
+    #   `if: :include_first_name?, or if: Proc.new { |user, options| options[:current_user] == user }).
+    #   The method, proc or string should return or evaluate to a true or false value.
+    # @option options [Symbol,Proc] :unless Specifies a method, proc or string
+    #   to call to determine if the field should be included (e.g.
+    #   `unless: :include_first_name?, or unless: Proc.new { |user, options| options[:current_user] != user }).
+    #   The method, proc or string should return or evaluate to a true or false value.
+    # @yield [object, options] The object and the options passed to render are
+    #   also yielded to the block.
     #
     # @example Specifying a user's first_name to be serialized.
     #   class UserBlueprint < Blueprinter::Base
@@ -64,21 +91,32 @@ module Blueprinter
     #
     # @example Passing a block to be evaluated as the value.
     #   class UserBlueprint < Blueprinter::Base
-    #     field :full_name {|obj| "#{obj.first_name} #{obj.last_name}"}
+    #     field :full_name do |object, options|
+    #       "options[:title_prefix] #{object.first_name} #{object.last_name}"
+    #     end
+    #     # other code
+    #   end
+    #
+    # @example Passing an if proc and unless method..
+    #   class UserBlueprint < Blueprinter::Base
+    #     def skip_first_name?(user, options)
+    #       user.first_name == options[:first_name]
+    #     end
+    #
+    #     field :first_name, unless: :skip_first_name?
+    #     field :last_name, if: ->(user, options) { user.first_name != options[:first_name] }
     #     # other code
     #   end
     #
     # @return [Field] A Field object
     def self.field(method, options = {}, &block)
-      options = if block_given?
-        {name: method, extractor: BlockExtractor, block: {method => block}}
-      else
-        {name: method, extractor: AutoExtractor}
-      end.merge(options)
-      current_view << Field.new(method,
-                                options[:name],
-                                options[:extractor],
-                                options)
+      current_view << Field.new(
+        method,
+        options.fetch(:name) { method },
+        options.fetch(:extractor) { AutoExtractor.new },
+        self,
+        options.merge(block: block),
+      )
     end
 
     # Specify an associated object to be included for serialization.
@@ -92,6 +130,8 @@ module Blueprinter
     #   JSON output.
     # @option options [Symbol] :view Specify the view to use or fall back to
     #   to the :default view.
+    # @yield [object, options] The object and the options passed to render are
+    #   also yielded to the block.
     #
     # @example Specifying an association
     #   class UserBlueprint < Blueprinter::Base
@@ -100,14 +140,22 @@ module Blueprinter
     #     # code
     #   end
     #
+    # @example Passing a block to be evaluated as the value.
+    #   class UserBlueprint < Blueprinter::Base
+    #     association :vehicles, blueprint: VehiclesBlueprint do |user, opts|
+    #       user.vehicles + opts[:additional_vehicles]
+    #     end
+    #   end
+    #
     # @return [Field] A Field object
-    def self.association(method, options = {})
+    def self.association(method, options = {}, &block)
       raise BlueprinterError, 'blueprint required' unless options[:blueprint]
-      name = options.delete(:name) || method
-      current_view << Field.new(method,
-                                       name,
-                                       AssociationExtractor,
-                                       options.merge(association: true))
+
+      field(
+        method,
+        options.merge(association: true, extractor: AssociationExtractor.new),
+        &block
+      )
     end
 
     # Generates a JSON formatted String.
@@ -127,8 +175,47 @@ module Blueprinter
     #
     # @return [String] JSON formatted String
     def self.render(object, options = {})
-      view_name = options.delete(:view) || :default
-      jsonify(prepare(object, view_name: view_name, local_options: options))
+      jsonify(prepare_for_render(object, options))
+    end
+
+    # Generates a hash.
+    # Takes a required object and an optional view.
+    #
+    # @param object [Object] the Object to serialize upon.
+    # @param options [Hash] the options hash which requires a :view. Any
+    #   additional key value pairs will be exposed during serialization.
+    # @option options [Symbol] :view Defaults to :default.
+    #   The view name that corresponds to the group of
+    #   fields to be serialized.
+    #
+    # @example Generating a hash with an extended view
+    #   post = Post.all
+    #   Blueprinter::Base.render_as_hash post, view: :extended
+    #   # => [{id:1, title: Hello},{id:2, title: My Day}]
+    #
+    # @return [Hash]
+    def self.render_as_hash(object, options= {})
+      prepare_for_render(object, options)
+    end
+
+    # Generates a JSONified hash.
+    # Takes a required object and an optional view.
+    #
+    # @param object [Object] the Object to serialize upon.
+    # @param options [Hash] the options hash which requires a :view. Any
+    #   additional key value pairs will be exposed during serialization.
+    # @option options [Symbol] :view Defaults to :default.
+    #   The view name that corresponds to the group of
+    #   fields to be serialized.
+    #
+    # @example Generating a hash with an extended view
+    #   post = Post.all
+    #   Blueprinter::Base.render_as_json post, view: :extended
+    #   # => [{"id" => "1", "title" => "Hello"},{"id" => "2", "title" => "My Day"}]
+    #
+    # @return [Hash]
+    def self.render_as_json(object, options= {})
+      prepare_for_render(object, options).as_json
     end
 
     # This is the magic method that converts complex objects into a simple hash
@@ -142,7 +229,6 @@ module Blueprinter
       unless view_collection.has_view? view_name
         raise BlueprinterError, "View '#{view_name}' is not defined"
       end
-      fields = view_collection.fields_for(view_name)
       prepared_object = include_associations(object, view_name: view_name)
       if array_like?(object)
         prepared_object.map do |obj|
@@ -172,13 +258,8 @@ module Blueprinter
     # @return [Array<Symbol>] an array of field names
     def self.fields(*field_names)
       field_names.each do |field_name|
-        current_view << Field.new(field_name, field_name, AutoExtractor)
+        field(field_name)
       end
-    end
-
-    # @api private
-    def self.associations(view_name = :default)
-      view_collection.fields_for(view_name).select { |f| f.options[:association] }
     end
 
     # Specify another view that should be mixed into the current view.
@@ -245,10 +326,21 @@ module Blueprinter
       @current_view = view_collection[:default]
     end
 
-    private
+    # Begin private class methods
+    def self.prepare_for_render(object, options)
+      view_name = options.delete(:view) || :default
+      prepare(object, view_name: view_name, local_options: options)
+    end
+    private_class_method :prepare_for_render
+
+    def self.inherited(subclass)
+      subclass.send(:view_collection).inherit(view_collection)
+    end
+    private_class_method :inherited
 
     def self.object_to_hash(object, view_name:, local_options:)
       view_collection.fields_for(view_name).each_with_object({}) do |field, hash|
+        next if field.skip?(object, local_options)
         hash[field.name] = field.extract(object, local_options)
       end
     end
@@ -273,7 +365,7 @@ module Blueprinter
     private_class_method :include_associations
 
     def self.jsonify(blob)
-      Blueprinter.configuration.generator.generate(blob)
+      Blueprinter.configuration.jsonify(blob)
     end
     private_class_method :jsonify
 
@@ -291,5 +383,10 @@ module Blueprinter
       object.is_a?(Array) || active_record_relation?(object)
     end
     private_class_method :array_like?
+
+    def self.associations(view_name = :default)
+      view_collection.fields_for(view_name).select { |f| f.options[:association] }
+    end
+    private_class_method :associations
   end
 end