+++ spec/benchmarks/active_record_ips_test.rb
@@ -1,9 +1,9 @@
-require_relative '../rails_test_helper'
-require_relative '../factories/model_factories.rb'
-require_relative 'benchmark_helper'
+require 'activerecord_helper'
+require 'benchmark_helper'
+require 'blueprinter/base'
 
 class Blueprinter::ActiveRecordIPSTest < Minitest::Test
-  include FactoryGirl::Syntax::Methods
+  include FactoryBot::Syntax::Methods
   include BenchmarkHelper
 
   def setup