+++ spec/dummy/app/models/user.rb
@@ -1,4 +1,0 @@
-class User < ApplicationRecord
-  attr_accessor :company, :description, :position
-  has_many :vehicles
-end