+++ spec/factories/model_factories.rb
@@ -1,14 +1,19 @@
-FactoryGirl.define do
+require 'factory_bot'
+
+FactoryBot.define do
   factory :user do
-    first_name 'Meg'
-    last_name  'Jones'
-    position 'Manager'
-    description 'A person'
-    company 'Procore'
+    first_name { 'Meg' }
+    last_name  { 'Jones' }
+    position { 'Manager' }
+    description { 'A person' }
+    company { 'Procore' }
+    birthday { Date.new(1994, 3, 4) }
+    deleted_at { nil }
+    active { false }
   end
 
   factory :vehicle do
-    make 'Super Car'
+    make { 'Super Car' }
     association :user
   end
 end