+++ spec/rails_test_helper.rb
@@ -1,4 +1,0 @@
-require File.expand_path("../../spec/dummy/config/environment.rb", __FILE__)
-Rails.env = 'test'
-ActiveRecord::Migrator.migrations_paths = [File.expand_path("../../spec/dummy/db/migrate", __FILE__)]
-load 'dummy/db/schema.rb'